Add Liquids-Handling Test
New test file `house-handling-liquids.a3sol` to test handling liquids
in containers by either the liquid's name or the container's name.

The test reveals a but with the `put_on` verb implemented on `liquid`,
where trying to use this verb with a liquid dosn't actually move the
liquid to the surface of the parameter.