Test Nested Clothing
Add a wearable item (a kippah) inside another (trousers pockets) to test
that the library correctly treats the former as a carried item.

TEST FAIL: At initialization time the library doesn't doesn't distinguish
between clothing items which are directly in the Hero from those that
are indirectly contained, resulting in the kippah being treated as worn.

ReRun all tests.