Reorganize Extras Contents
In view of upcoming AsciiDoc documentation (StdLib Manual and Cookbook)
start restructuring the `extras_src/` and `extras/` folders, so that the
same script will be able to process every documentation source and target
folders symmetrically, using the same functions.

- New "Clothing Guide" locations:
  -  sources: `/extras_src/tutorials/` (both AsciiDoc and Alan files).
  -  destination: `/extras/tutorials/`
- Temporary UTF-8 folder: `/extras_src/utf8/`

Note that now the AsciiDoc sources and the Alan adventures and commands
scripts are all kept together in the same folder.