StdLib Manual: Chapters IDs
Add a custom ID to every chapter with form `[[ch1]]`, `[[ch2]]`, ecc.,
to simplify xrefs and preserve them if titles change.

Start restoring lost text in Ch.1 -- in the early stages of the porting
work I didn't realize that chunks of text went missing, so with the
first chapters I didn't double check with the PDF, but when later on I
started fixing them on the spot.