Fix Manual Example and Externalized Code
Fix the source of short example 6, which still used the old modules
names. In Ch.12, fix the `include::` directive to use the new name of
`lib_messages.i` --> `lib_messages_runtime.i`, and change the length
of the ALAN code delimiters to avoid conflicts with the 80-dashes
comment line in the library source.

Rebuild all documents.