StdLib Man Quick Builder
Create `extras_src/manual/html-build.sh`, a script to quickly update
the StdLib Manual without processing all the ALAN source examples, in
order to allow quick previews of changes to the ADoc sources.

The script was quickly setup in a rather hackish way, and will soon be
replaced with a better version (all toolachain scripts will).