StdLib Man: Revise Intro
Move the ALAN-only bare-bones adventure example from "Ch. 2. Importing
the Library" to the new section "1. Who Needs a Library?" under the
main "Introduction".

In "1. Who Needs a Library?", write an introduction to ALAN's design
philosophy of providing a bare-bones and locale agnostic language, and
how this compares to other IF systems, and how the StdLib aims to provide
a basic library for authors who don't want to write everything from the
ground up.