Library Classes Diagram
Create a diagram of the StdLib classes, using the Dia diagram creator:

    http://dia-installer.de

Store the Dia project in the new `assets_src/images/dia/` folder, and
add a script to convert all Dia projects to SVG, optimize them via SVGO,
and then deploy them to `assets/images/`.

Since the generated SVGs are copied to `assets/images/`, add a rule to
`.gitignore` to exclude all SVG files in the Dia sources folder.