Fix Unlit Location Bug
In `lib_locations.i` add a DESCRIPTION CHECK to EVERY LOCATION to ensure
that if the location is `NOT lit` the `dark_loc_desc` message will be
show instead (See #96).

There might be more fixes to be done in this regard, e.g. covering NPCs
that enter the location and whether to execute a LOOK statement or not
(See discussion at #96 for more info).