Fix deployAlan Func
In `_build-funcs.sh`, fix the SED RegEx in `deployAlan` function which
was not matching ADoc tags containing hyphens. Now the RegEx only looks
for the opening `tag::` and `end::` markers, to be on the safe side.

Update the deployed examples in `extras/manual/` that contained leaked
ADoc region tags, substituting them with their cleaned-up versions, free
from the spurious region tags.