Manual: Ch 14. Run-Time Messages
Rebuild the StdLib Manual to mirror the updated `lib_messages_runtime.i`
module, which is imported in Ch 14. Also, polish the introductory text.

BUG: Syntax highlighting of the updated source file is breaking up after
the `CANT0` message definition, where the following string is not being
ended properly. There might be an issue with the current ALAN syntax for
Highlight. Should look into it (see #102).