Fix Meta-Verbs Test Issues
In `tests/misc/` rename `meta-verbs.a3sol` to `meta-verbs-tests.a3sol`
because the solution file contains a `script on` command that causes
the creation of a `meta-verbs.a3t` file, which would interfere with
out generated transcript when we'll switch to the new `.at3` and `.ats`
extensions (see #128).

This change is due to the fact that now ARun no longer timestamps the
transcript when invoked with the `-r` switch, and because it now uses
the `.a3t` extension for transcripts (see alan-if/alan#2).

The new solution filename will now ensure that our test transcript is
named `meta-verbs-tests.a3log`, which will not be overridden when we
adopt the new `.a3t` extension for transcripts.

Add `/tests/misc/meta-verbs.a3t` to Git ignore rules.

Add a warning note about this in `meta-verbs.alan`.