Switch to ALAN Beta8 and UTF-8
This commit adopts the latest Beta8 SDK which supports UTF-8 encoding of
ALAN files (fixes #131). All ALAN files are transcoded for ISO/Latin-1:

|   ext.  |  encoding |                 file type                  |
|---------|-----------|--------------------------------------------|
| `.a3s`  | UTF-8 BOM | "solution" files (aka "commands scripts"). |
| `.a3t`  | UTF-8     | game transcripts.                          |
| `.alan` | UTF-8 BOM | ALAN source files.                         |
| `.i`    | UTF-8 BOM | ALAN included source files.                |

In AsciiDoc sources the `encoding=iso-8859-1` option is removed from all
`include::` directives dealing with ALAN sources and transcript.

All test suite batch scripts are deleted and replaced with a Rake task,
to prevent the BOM from the solution file to end up in the generated
transcripts (see alan-if/alan#32).

README files and the StdLib CHANGELOG are updated accordingly.

This commit is the first step of a series of big changes leading to the
adoption by the StdLib repository of new standards and tools that are
being shared by various ALAN repositories (see #133).