Polish Clothing Dev Notes
Revise the `CLOTHING_NEW.md` and `CLOTHING_DEV.md` temporary
documents by mentioning the new `wearer` attribute in the
former, and polishing the text of both docs.

Also add a clear notice to both regarding the fact that they
might not accurately mirror the current state of the library
codebase, and that they are being kept mostly as references
from which to copy and paste contents for the creation of
the final document that will introduce the new clothing
system (in AsciiDoc format, most probably).